+++ apps/web/src/constants/index.ts
@@ -3,6 +3,7 @@ export * from './header.constants';
 export * from './landing.constants';
 export * from './mapPage.constants';
 export * from './mypageNav.constants';
+export * from './otherspagetap.constants';
 export * from './sign_up.constants';
 export * from './sign_up.constants';
 export * from './stepnavbar.constants';
