+++ apps/web/src/store/index.ts
@@ -1,5 +1,6 @@
 export * from './floatingMenu.atoms';
 export * from './map.atoms';
+export * from './otherStoryTab.atom';
 export * from './selectedDateIdAtom.atom';
 export * from './step1Title.atoms';
 export * from './stepSumup.atoms';
